try to fix lazy exception

+++ tm-server/src/main/java/ru/girfanov/tm/endpoint/ProjectEndPoint.java
@@ -3,10 +3,12 @@ package ru.girfanov.tm.endpoint;
 import lombok.NoArgsConstructor;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
+import org.jetbrains.annotations.Nullable;
 import ru.girfanov.tm.api.service.IProjectService;
 import ru.girfanov.tm.api.service.ISessionService;
 import ru.girfanov.tm.entity.Project;
 import ru.girfanov.tm.entity.Session;
+import ru.girfanov.tm.exception.UserNotFoundException;
 import ru.girfanov.tm.exception.WrongSessionException;
 
 import javax.jws.WebMethod;
@@ -25,70 +27,69 @@ public final class ProjectEndPoint {
     @WebMethod
     public void persistProject(@WebParam(name = "session") final Session session, @WebParam(name = "project") final Project project) {
         try {
-            sessionService.existSession(session);
-        } catch (WrongSessionException e) {
+            if(sessionService.existSession(session)) projectService.persist(session.getUser().getId(), project);
+        } catch (WrongSessionException | UserNotFoundException e) {
             System.out.println(e.getMessage());
         }
-        projectService.persist(session.getUser().getId(), project);
     }
 
     @WebMethod
     public void mergeProject(@WebParam(name = "session") final Session session, @WebParam(name = "project") final Project project) {
         try {
-            sessionService.existSession(session);
-        } catch (WrongSessionException e) {
+            if(sessionService.existSession(session)) projectService.merge(session.getUser().getId(), project);
+        } catch (WrongSessionException | UserNotFoundException e) {
             System.out.println(e.getMessage());
         }
-        projectService.merge(session.getUser().getId(), project);
     }
 
     @WebMethod
     public void removeProject(@WebParam(name = "session") final Session session, @WebParam(name = "project") final Project project) {
         try {
-            sessionService.existSession(session);
-        } catch (WrongSessionException e) {
+            if(sessionService.existSession(session)) projectService.remove(session.getUser().getId(), project);
+        } catch (WrongSessionException | UserNotFoundException e) {
             System.out.println(e.getMessage());
         }
-        projectService.remove(session.getUser().getId(), project);
     }
 
     @WebMethod
     public void removeAllProjects(@WebParam(name = "session") final Session session) {
         try {
-            sessionService.existSession(session);
-        } catch (WrongSessionException e) {
+            if(sessionService.existSession(session)) projectService.removeAllByUserId(session.getUser().getId());
+        } catch (WrongSessionException | UserNotFoundException e) {
             System.out.println(e.getMessage());
         }
-        projectService.removeAllByUserId(session.getUser().getId());
     }
 
+    @Nullable
     @WebMethod
     public Project findOneProject(@WebParam(name = "session") final Session session, @WebParam(name = "projectUuid") final String projectUuid) {
         try {
-            sessionService.existSession(session);
-        } catch (WrongSessionException e) {
+            if(sessionService.existSession(session)) return projectService.findOne(session.getUser().getId(), projectUuid);
+        } catch (WrongSessionException | UserNotFoundException e) {
             System.out.println(e.getMessage());
         }
-        return projectService.findOne(session.getUser().getId(), projectUuid);
+        return null;
     }
 
+    @Nullable
     @WebMethod
     public List<Project> findAllProjects(@WebParam(name = "session") final Session session) {
         try {
-            sessionService.existSession(session);
-        } catch (WrongSessionException e) {
+            if(sessionService.existSession(session)) return projectService.findAllByUserId(session.getUser().getId());
+        } catch (WrongSessionException | UserNotFoundException e) {
             System.out.println(e.getMessage());
         }
-        return projectService.findAllByUserId(session.getUser().getId());
+        return null;
     }
 
+    @Nullable
     @WebMethod
     public List<Project> findAllProjectsSortedByValue(@WebParam(name = "session") final Session session, @WebParam(name = "value") final String value) {
         try {
-            sessionService.existSession(session);
-        } catch (WrongSessionException e) {
+            if(sessionService.existSession(session)) return projectService.findAllSortedByValue(session.getUser().getId(), value);
+        } catch (WrongSessionException | UserNotFoundException e) {
             System.out.println(e.getMessage());
         }
-        return projectService.findAllSortedByValue(session.getUser().getId(), value);
+        return null;
     }
 }
