+++ tm-server/src/main/java/ru/girfanov/tm/endpoint/UserEndPoint.java
@@ -3,10 +3,12 @@ package ru.girfanov.tm.endpoint;
 import lombok.NoArgsConstructor;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
+import org.jetbrains.annotations.Nullable;
 import ru.girfanov.tm.api.service.ISessionService;
 import ru.girfanov.tm.api.service.IUserService;
 import ru.girfanov.tm.entity.Session;
 import ru.girfanov.tm.entity.User;
+import ru.girfanov.tm.exception.UserNotFoundException;
 import ru.girfanov.tm.exception.WrongSessionException;
 
 import javax.jws.WebMethod;
@@ -30,41 +32,41 @@ public final class UserEndPoint {
     @WebMethod
     public void mergeUser(@WebParam(name = "session") final Session session, @WebParam(name = "user") final User user) {
         try {
-            sessionService.existSession(session);
+            if(sessionService.existSession(session)) userService.merge(user);
         } catch (WrongSessionException e) {
             System.out.println(e.getMessage());
         }
-        userService.merge(user);
     }
 
     @WebMethod
     public void removeUser(@WebParam(name = "session") final Session session, @WebParam(name = "user") final User user) {
         try {
-            sessionService.existSession(session);
+            if(sessionService.existSession(session)) userService.remove(user);
         } catch (WrongSessionException e) {
             System.out.println(e.getMessage());
         }
-        userService.remove(user);
     }
 
+    @Nullable
     @WebMethod
     public User findOneUser(@WebParam(name = "session") final Session session, @WebParam(name = "userUuid") final String userId) {
         try {
-            sessionService.existSession(session);
-        } catch (WrongSessionException e) {
+            if(sessionService.existSession(session)) return userService.findOne(userId);
+        } catch (WrongSessionException | UserNotFoundException e) {
             System.out.println(e.getMessage());
         }
-        return userService.findOne(userId);
+        return null;
     }
 
+    @Nullable
     @WebMethod
     public List<User> findAllUsers(@WebParam(name = "session") final Session session) {
         try {
-            sessionService.existSession(session);
+            if(sessionService.existSession(session)) return userService.findAll();
         } catch (WrongSessionException e) {
             System.out.println(e.getMessage());
         }
-        return userService.findAll();
+        return null;
     }
 
     @WebMethod //for serialization
@@ -72,8 +74,14 @@ public final class UserEndPoint {
         return userService.findAll();
     }
 
+    @Nullable
     @WebMethod
-    public User findOneUserByLoginAndPassword(@WebParam(name = "login") final String login, @WebParam(name = "password") final String password) {
-        return userService.findOneByLoginAndPassword(login, password);
+    public User findOneUserByLogin(@WebParam(name = "login") final String login) {
+        try {
+            return userService.findOneByLogin(login);
+        } catch (UserNotFoundException e) {
+            System.out.println(e.getMessage());
+        }
+        return null;
     }
 }
