+++ tm-server/src/main/java/ru/girfanov/tm/service/ProjectService.java
@@ -22,76 +22,64 @@ public final class ProjectService implements IProjectService {
     @NonNull private EntityManagerFactory entityManagerFactory;
 
     @Override
-    public void persist(@NotNull final String userId, @NotNull final Project project) {
+    public void persist(@NotNull final String userId, @NotNull final Project project) throws UserNotFoundException {
         if(userId.isEmpty()) { return; }
         final EntityManager em = entityManagerFactory.createEntityManager();
         final ProjectRepository projectRepository = new ProjectRepository(em);
         final UserRepository userRepository = new UserRepository(em);
         try {
             em.getTransaction().begin();
-            userRepository.findOne(userId);
+            if(userRepository.findOne(userId) == null) throw new UserNotFoundException("user not found");
             projectRepository.persist(project);
             em.getTransaction().commit();
-        } catch (UserNotFoundException e) {
-            em.getTransaction().rollback();
-            System.out.println(e.getMessage());
         } finally {
             em.close();
         }
     }
 
     @Override
-    public void merge(@NotNull final String userId, @NotNull final Project project) {
+    public void merge(@NotNull final String userId, @NotNull final Project project) throws UserNotFoundException {
         if(userId.isEmpty()) { return; }
         final EntityManager em = entityManagerFactory.createEntityManager();
         final ProjectRepository projectRepository = new ProjectRepository(em);
         final UserRepository userRepository = new UserRepository(em);
         try {
             em.getTransaction().begin();
-            userRepository.findOne(userId);
+            if(userRepository.findOne(userId) == null) throw new UserNotFoundException("user not found");
             projectRepository.merge(project);
             em.getTransaction().commit();
-        } catch (UserNotFoundException e) {
-            em.getTransaction().rollback();
-            System.out.println(e.getMessage());
         } finally {
             em.close();
         }
     }
 
     @Override
-    public void remove(@NotNull final String userId, @NotNull final Project project) {
+    public void remove(@NotNull final String userId, @NotNull final Project project) throws UserNotFoundException {
         if(userId.isEmpty()) { return; }
         final EntityManager em = entityManagerFactory.createEntityManager();
         final ProjectRepository projectRepository = new ProjectRepository(em);
         final UserRepository userRepository = new UserRepository(em);
         try {
             em.getTransaction().begin();
-            userRepository.findOne(userId);
+            if(userRepository.findOne(userId) == null) throw new UserNotFoundException("user not found");
             projectRepository.remove(project);
             em.getTransaction().commit();
-        } catch (UserNotFoundException e) {
-            em.getTransaction().rollback();
-            System.out.println(e.getMessage());
         } finally {
             em.close();
         }
     }
 
     @Override
-    public void removeAllByUserId(@NotNull final String userId) {
+    public void removeAllByUserId(@NotNull final String userId) throws UserNotFoundException {
         if(userId.isEmpty()) { return; }
         final EntityManager em = entityManagerFactory.createEntityManager();
         final ProjectRepository projectRepository = new ProjectRepository(em);
         final UserRepository userRepository = new UserRepository(em);
         try {
             em.getTransaction().begin();
-            userRepository.findOne(userId);
+            if(userRepository.findOne(userId) == null) throw new UserNotFoundException("user not found");
             projectRepository.removeAllByUserId(userId);
             em.getTransaction().commit();
-        } catch (UserNotFoundException e) {
-            em.getTransaction().rollback();
-            System.out.println(e.getMessage());
         } finally {
             em.close();
         }
@@ -99,17 +87,15 @@ public final class ProjectService implements IProjectService {
 
     @Nullable
     @Override
-    public Project findOne(@NotNull final String userId, @NotNull final String projectId) {
+    public Project findOne(@NotNull final String userId, @NotNull final String projectId) throws UserNotFoundException {
         if (userId.isEmpty() || projectId.isEmpty()) { return null; }
         Project project = null;
         final EntityManager em = entityManagerFactory.createEntityManager();
         final ProjectRepository projectRepository = new ProjectRepository(em);
         final UserRepository userRepository = new UserRepository(em);
         try {
-            userRepository.findOne(userId);
+            if(userRepository.findOne(userId) == null) throw new UserNotFoundException("user not found");
             project = projectRepository.findOne(userId, projectId);
-        } catch (UserNotFoundException e) {
-            System.out.println(e.getMessage());
         } finally {
             em.close();
         }
@@ -118,17 +104,15 @@ public final class ProjectService implements IProjectService {
 
     @Nullable
     @Override
-    public List<Project> findAllByUserId(@NotNull final String userId) {
+    public List<Project> findAllByUserId(@NotNull final String userId) throws UserNotFoundException {
         if(userId.isEmpty()) { return null; }
         List<Project> projects = null;
         final EntityManager em = entityManagerFactory.createEntityManager();
         final ProjectRepository projectRepository = new ProjectRepository(em);
         final UserRepository userRepository = new UserRepository(em);
         try {
-            userRepository.findOne(userId);
+            if(userRepository.findOne(userId) == null) throw new UserNotFoundException("user not found");
             projects = projectRepository.findAllByUserId(userId);
-        } catch (UserNotFoundException e) {
-            System.out.println(e.getMessage());
         } finally {
             em.close();
         }
@@ -147,17 +131,15 @@ public final class ProjectService implements IProjectService {
 
     @Nullable
     @Override
-    public List<Project> findAllSortedByValue(@NotNull final String userId, @NotNull final String value) {
+    public List<Project> findAllSortedByValue(@NotNull final String userId, @NotNull final String value) throws UserNotFoundException {
         if (userId.isEmpty() || value.isEmpty()) { return null; }
         List<Project> projects = null;
         final EntityManager em = entityManagerFactory.createEntityManager();
         final ProjectRepository projectRepository = new ProjectRepository(em);
         final UserRepository userRepository = new UserRepository(em);
         try {
-            userRepository.findOne(userId);
+            if(userRepository.findOne(userId) == null) throw new UserNotFoundException("user not found");
             projects = projectRepository.findAllSortedByValue(userId, value);
-        } catch (UserNotFoundException e) {
-            System.out.println(e.getMessage());
         } finally {
             em.close();
         }
